+++ tools/reso-reference-server/compliance/entrypoint-core.sh
@@ -28,6 +28,7 @@ for script in /tmp/resoscripts/*.resoscript; do
     -DpathToRESOScript="$script" \
     -DuseStringEnums=true \
     -DuseCollections=true \
+    -DshowResponses=true \
     || FAILED=1
 done